search 2.1

+++ app/src/main/java/com/example/banksampah/model/UserHistori.kt
@@ -2,7 +2,7 @@ package com.example.banksampah.model
 
 data class UserHistori (
     var hargaSubKategori: Long ?= 0,
-    var jumlah: Long?=null,
+    var jumlah: Double?=null,
     var kategori:String?=null,
     var namaBank:String?=null,
     var namaNasabah:String?=null,
